feat: breakup pages into component templates
