+++ src/components/partners/partners.jsx
@@ -6,7 +6,7 @@ import React from "react";
 const Partners = () => {
   return (
     <Container className="text-center">
-      <div className=" px-8 md:py-14 py-10 rounded-lg  border border-blue bg-white bg-opacity-5 ">
+      <div className=" px-8 md:py-14 py-10 rounded-lg overflow-scroll  border border-blue bg-white bg-opacity-5 ">
         {partners.map(partner => {
           return <Partner key={partner.name} {...partner} />;
         })}
@@ -17,7 +17,7 @@ const Partners = () => {
 
 const Partner = ({ Logo, href }) => {
   return (
-    <NLink href={href}>
+    <NLink href={href} className="inline-block">
       <Logo height="28px" className="inline-block mx-3" />
     </NLink>
   );
