+++ src/templates/about-us/index.js
@@ -1,0 +1,2 @@
+export * from "./header";
+export * from "./team";
